move olMap related files  to new folder "olMap"
