+++ src/modules/map/components/olMap/index.js
@@ -1,0 +1,4 @@
+import { OlMap } from './OlMap';
+if (!window.customElements.get(OlMap.tag)) {
+	window.customElements.define(OlMap.tag, OlMap);
+}
