+++ src/modules/map/index.js
@@ -2,7 +2,4 @@ import './i18n';
 import './components/zoomButtons';
 import './components/layerManager';
 import './components/infoButton';
-import { OlMap } from './components/OlMap';
-if (!window.customElements.get(OlMap.tag)) {
-	window.customElements.define(OlMap.tag, OlMap);
-}
+import './components/olMap';
